fix: pass size to Column and render Columns in RN
