+++ morph/react-dom/block.js
@@ -1,6 +1,6 @@
 import * as BlockAddTestIdProp from '../react/block-add-test-id-prop.js'
 import * as BlockCapture from './block-capture.js'
-import * as BlockColumn from './block-column.js'
+import * as BlockColumn from '../react/block-column.js'
 import * as BlockExplicitChildren from '../react/block-explicit-children.js'
 import * as BlockGoTo from './block-go-to.js'
 import * as BlockGroup from '../react/block-group.js'
@@ -12,6 +12,7 @@ import * as BlockProperties from './block-properties.js'
 import * as BlockProxy from '../react/block-proxy.js'
 import * as BlockRoute from '../react/block-route.js'
 import * as BlockSetTestId from '../react/block-set-test-id.js'
+import * as BlockTable from '../react/block-table.js'
 import * as BlockTeleport from '../react/block-teleport.js'
 
 export const enter = [
@@ -29,9 +30,11 @@ export const enter = [
   BlockProperties.enter,
   BlockGroup.enter,
   BlockList.enter,
+  BlockTable.enter,
 ]
 
 export const leave = [
+  BlockTable.leave,
   BlockList.leave,
   BlockExplicitChildren.leave,
   BlockName.leave,
