+++ morph/react-native/block.js
@@ -1,5 +1,6 @@
 import * as BlockBackgroundImage from './block-background-image.js'
 import * as BlockCapture from './block-capture.js'
+import * as BlockColumn from '../react/block-column.js'
 import * as BlockExplicitChildren from '../react/block-explicit-children.js'
 // import * as BlockGoTo from './block-go-to.js'
 import * as BlockInList from '../react/block-in-list.js'
@@ -11,6 +12,7 @@ import * as BlockProperties from './block-properties.js'
 import * as BlockProxy from '../react/block-proxy.js'
 import * as BlockRoute from '../react/block-route.js'
 import * as BlockSetTestId from '../react/block-set-test-id.js'
+import * as BlockTable from '../react/block-table.js'
 import * as BlockTeleport from '../react/block-teleport.js'
 import * as BlockAddTestIdProp from '../react/block-add-test-id-prop.js'
 import * as BlockOffWhen from '../react/block-off-when.js'
@@ -23,6 +25,7 @@ export const enter = [
   BlockRoute.enter,
   BlockWrap.enter,
   BlockName.enter,
+  BlockColumn.enter,
   BlockSvg.enter,
   BlockCapture.enter,
   BlockBackgroundImage.enter,
@@ -33,9 +36,11 @@ export const enter = [
   BlockProperties.enter,
   BlockGroup.enter,
   BlockList.enter,
+  BlockTable.enter,
 ]
 
 export const leave = [
+  BlockTable.leave,
   BlockList.leave,
   BlockExplicitChildren.leave,
   BlockName.leave,
