+++ morph/react/block-table.js
@@ -1,0 +1,13 @@
+import { isTable } from '../utils.js'
+
+export function enter(node, parent, state) {
+  if (!isTable(node)) return
+
+  state.render.push('{size => (<React.Fragment>')
+}
+
+export function leave(node, parent, state) {
+  if (!isTable(node)) return
+
+  state.render.push('</React.Fragment>)}')
+}
